code: Add update_bandwidth() API call
The configured bandwidth is node-wide, i.e. all the shards are
enforced to stay below the limit altogether. However, since the
throttling engine is per-IO-group the only possible option is
to evenly split the limit between groups giving each the 1/Nth
of the original value.

To make it work like this the API call submits the request to
all shards giving each the 1/Nth of the original value, then
only the shards that "own" the group (i.e. -- allocated it) set
the group limit.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>