shared_token_bucket: Allow changing rate
So far the configured rate was very low to care about any overflows.
With the ability to change the rate there appears a (tiny but stil)
risk that converting time -> tokens overflows after multiplication of
rate by the duration. So other than de-constifying the replenisher
rate, this patch also adds some (very loose) restrictions on the
maximum allowed rate.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>