reactor: Always limit configuration with active cpumask
When seastar app is run under the e.g. taskset shell command it
effectively ignores the enforced cpumask and pins its thread to any
core (well, to whatever core is allowed by current cgroup which is
pretty much often is "any").

Fix this by assuming that the --cpus is always present, but the empty
value means accepting current task's cpumask.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>