sharded: De-duplicate map-reduce overloads
There are two (well, three) of them -- one for Service::*member
invocation, the other one with generic Func&& argument. And there's
a const version of the latter (only).

This patch merges both with the help of std::invoke removing
duplicate code and at the same time adding the const version for
the class method invocation.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>