rpc: Send messages instantly
Sending a message is serialized with the help of a background fiber.
There's a std::list of outgoing entries that is filled in send(), then a
condition variable is signalled and the list is harvested by the fiber.
This introduces a wake-up latency on both send and ack paths which can
generate noticeable RPC slow-down.

The fix is in serializing messages with the help of a plain sempahore.
The entry in question is then kept in the continuation chain and when
the semaphore allows the entry is put on the wire. Waiting the semaphore
is abortable so that canceled requests resolve their futures without
waiting for the other pending messaging to resolve.

While sitting in the outgoing queue the request can be cancelled,
explicitly or by a timeout. To keep this support what sits on the
continuation is the std::optional<entry> that's .reset() upon cancel.
The sending continuation then just checks if the entry is alive and
resolves into a ready future if now.

Stopping this brand new background queue is done by breaking the
semaphore and waiting for any of its holders (just one) to release it.

The outgoing queue length is now the semaphore wait-queue length.

Suspend and resume is in locking/unlocking the semaphore respectively.

The old promise that sat on the entry and was waited on by the caller is
no longer needed, the immediate future can be used instead.

Indentation is deliberately left broken in some places.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>