semaphore: Add break_and_wait() methods
The method in question breaks the semaphore as the .broken() would do
and returns a future that will be resolved once semaphore is signalled
back to the specified amount of units.

This is done by placing a regular waiter on the waitlist (or resolving
it instantly if the semaphore is not locked) and not resetting the
counter to zero. Callers of .signal() would notify the waiter as they
normally do.

Re-entering .wait() after .break_and_wait() is _still_ no possible as
the semaphore would not have enough units or its wait list would not be
empty.

Tests included.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>