rpc: Drop outgoing_queue_type
It's needed to decide if the 8 bytes timeout value should be at the
message head or not. Now this is known when the outgoing queue loop
starts, but next patches will remove this loop and this decision will be
taken upon each .send() invocation.

The fix is in the explicit _propagate_timeout boolean that's false by
default and is set to true when connection decides to.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>