rpc, test: Fixup cancelling case
The test case tries to cancel a request after it was called, but before
it was sent. The operator() thing doesn't guarantee that it will delay
sending, it just works today because send() offloads the actual sending
to the background send_loop fiber.

Fix this by introducing suspend/resume calls for connection. For now
they are no-op just because they will be re-implemented by the next
patch.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>