semaphore: Add get_units_abortable
The helper tries to lock a semaphore with an abort-source, but returns
valid future if the abort-source fires. In that case the resolved units
are "empty" (called disengaged in this patch) -- contains no units and
doesn't point to a semaphore.

Useful for cases when aborting a semaphore is benign situation that
needs to be handled without resolving into any exceptions.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>