rpc: Send messages instantly
Sending a message is serialized with the help of a background fiber.
There's a std::list of outgoing entries that is filled in send(), then a
condition variable is signalled and the list is harvested by the fiber.
This introduces a wake-up latency on both send and ack paths which can
generate noticeable RPC slow-down.

The fix is in serializing messages through a "classical" continuation
chain. For the entries to be cancellable they still need to be linked
in a list so that a cancelled entry can withdraw itself from the chain.

Cancelling (and withdrawing from the list) of the first entry is not
possible, because the head of the outgoing queue is likely sittiing in
the fd.writable()/fd.flush()/kernel and cannot be aborted.

Non-front entries can be aborted and the respective continuation chain
is fired out of order. Whether the entry is cancelled or not is then
checked in the first chain lambda. Since the entry should stay alive
in memory after it's erased from the queue, the std::list is replaced
with boost intrusive one.

The old promise that sat on the entry and was waited on by the caller is
no longer needed, the immediate future can be used instead.

Indentation is deliberately left broken in some places.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>