io_request: Add .split(size_t max_length) method
The method splits the request into a bunch of other requests all (but
the last one) being exactly the max_length bytes. This functionality
will be needed by the io-queue to split oversized requests.

Unit test included. Splitting the request backed by a buffer is simple
-- just need to validate three cases, when the max_length is larger,
equal and smaller than the buffer itself. Splitting the request carrying
iovecs on board is done in a fuzzy manner -- random iovecs are splitted
into ranom max_length parts, then the underlying memory is filled via
both -- the original one and the target set, then the buffer is checked
to contain the expected values.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>