output_stream: Make stream::flush() resolve into batch-flush result
When flushes are batched the .flush() method resolves immediatelly which
is probably not very good. Consider this flow

    while (true) {
        co_await client_stream.write();
        co_await client_stream.flush();
        co_await client_stream.recv();
        co_await handle();
    }

if .flush() is batched and the underlying "real" flush resolves with
exception the whole thing gets stuck because

    .flush() resolves without exception as was shown above
    .recv() hangs waiting for the response that may never arrive, because
            request sending failed
    .handle() won't get called
    .write() on next loop iteration may notice the exception, but it won't
             be called either

Fix this by making stream::flush() wait for the batch flush to take
place and report back its future whatever it is.

fixes: #1150

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>