output_stream: Introduce on-batch-flush-error callback
When output_strea::flush() is called with batch flushes ON the flush()
itself resolves immediately with ready future. The flush itself happens
later when a dedicated poller gets to it. If that real flush resolves
with exception the poller just keeps the exception on the stream and
does nothing more.

The described behavior can dead-lock if the user operates in the

    while (!aborted()) {
        co_await ostream.send()
        co_await ostream.flush()
        co_await istream.recv()
        co_await handle()
    }

manner. Even if ostream real flush fails some time later the whole loop
is not aware of it as it's sitting in another stream.

This patch adds a callback to batch flush context that's called when
flush fails in the background. The callback is to be provided by the
stream creator, if not the batch-flushes are left OFF. This slightly
changes the API -- all current callers of connected_socket::output() now
get the synchronous flushing.

refs: #1150

(indentation is deliberately left broken)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>