file: Make IO methods use default priority class explicitly
All existing file IO calls accept the io_priority_class argument which
defaults to default_priority_class() if omitted. The plan is to
deprecate the ability to pass explicit priority class in this call in
faror of using priority class derived from current_scheduling_group().

To make the deprecation possible all the file methods are explicitly
overloaded to accept priorty class argument and not to (and use
default). The former set would later be deprecated, and the latter one
will change its semantics (in a controllable way).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>