rpc: Collect stats into stat_group's
There's the need to report RPC stats as metrics and there are two
extreme ways on how to do it.

First is to make each connection export metrics about itself. It's
possible, but since typically there are A LOT of connections, this would
explode the reported metrics list. Also, as connections appear and die
dynamically, it will be inconvenient to collect them on the client-side.

Second option is to sum up all the connections' counters into one place.
That's not great either, e.g. in multi-datacenter set-ups latency
metrics for intra-DC communications would be much better than those for
cross-DC ones. Mixing them together is producing some useless "average".
Another reason is that scylla uses different connections for RPC of
different "importance" to avoid head-of-line blocking problem. It's much
more informative to get separate sent/recvd/errors stats.

Having said that, the proposal is to introduce a stats_group abstraction
that would accumulate statistics for sockets that are attached to it and
it's up to the caller how to groups the connections.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>