rpc: Add metrics to stat_groups
Expose all the existing rpc::stats counters and the current number of
attached connections gauge.

The stats group explicitly lists the attached sockets, so when reporting
the metrics the group code walks the list of connections and accumulates
those in one "snapshot" that's reported later. That's not to ++ counters
in two different cache-lines as was described in previous patch.

Not to do the walk per-counter, the scrape notification is used.

To handle the case when a connection dies and can no longer be walked
for metrics, dying connections accumulate their counters on group's
"dead" store.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>