core: Suspend stack printing if caught in the middle of exception
Stall detector and SEGV/ABRT handlers try to print a backtrace when
caught. We've seen that sometimes collecting a backtrace steps on
exception stack unwinding and crashes. That's not a big deal for
SEGV/ABRT handling, since app is crashing anwyay, but for stall detector
this is troublesome.

Add a reactor option to avoid backtrace collection if the detector is
signalled in the middle of exception handling.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>