io_queue: Introduce IO controller mode of dispatching
In this mode, the existing fair capacity accounting is not performed.
Instead, each "tick" the whole accumulated queue is flushed into disk.
The only reason to stop this flushing is "unplugging" the class which,
in turn, happens when a class hists its bandwidth or IOPS limit.

The mode is activated by `--io-controller=true` seastar option.

The dedicated mode is, strictly speaking, not necessary. The existing
fair-queue loop would work with unlimited io-properties just as good.
However, we're going to keep using the real io-properties numbers to
configure the controller, so instead of tuning the configuration phase,
let's better be explicit and just plug the dispatch loop for controller
mode.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>