io_queue: Split disk bandwidth into per-class bandwidth limits
When a class is configured using update_bandiwdth(class, value) method,
the controller mode changes the way it's applied.

In this mode, the configured bandwidth is considered to be a "demand"
for the bandiwdth the class requests. Given demands for all classes, we
can iterate over them and calculate bandwidth limits so that for every
class "c"

   disk_bandwidth - sum(bandwidth_limit[i], i!=c) >= demand[i]

i.e. -- if some class demands for some bandwidth, all other classes are
configured to consume not more than total minus that demand. And this
stays simultaneously true for all registered classes.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>