fair_queue: Move replenisher kick to dispatch caller
When fair_queue dispatching starts, it first tries to replenish capacity
tokens to outgoing throttler. Move this to fair_queue caller, namely --
to the io_queue. For that, the io_queue::stream gains reference on the
io_throttler. It's the same throttler instance that the corresponding
fair_queue uses now.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>