fair_queue: Swap class push back and dispatch
When a class is dispatched from it should be placed back into fair_queue
priority queue for further dispatching with updated accumulator.
However, the class can be unplugged by the dispatching callback, so the
push back is done after the callback.

These two can in fact be swapped. Even if a class is pushed back before
dispatch and is unplugged, the next dispatch loop will notice the unplug
bit and drop the class anyway.

Doing push-back and dispatch in that order makes further patching
possible.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>