fair_queue: Split dispatch_requests() into top() and pop_front()
Make the API to maniupulate fair_queue in a priority-queue-like manner.
Add two methods -- top() returns the next entry to dispatch and
pop_front() withdraws it from the queue and rebalances classes.

This will allow to get rid of the "dispatch" callable passed as an
argument to the dispatch_requests() and eventually implemente IO-queue
level throttling.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>