fair_queue: Open-code dispatch_requests() into users
Now the dispatch_requests() is short and simple enough for callers to
implement it on their own. All the more so different users (io_queue and
tests) have different idea of how to grab capacity for dispatch. This
makes it possible to move outgoing capacity grabbing to IO-queue and
makes tests simpler.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>