test: Add test for socket shutdown-read sanity
The test checks how shutting down read for a connected socket affects
the pending readers from the same socket.

First part starts a background read from an empty socket and shuts down
its input. The expectation is read() resolves in an empty buffer,
indicating EOF.

Second part does the same but when the socket has some data in it before
shutwodn is called. The expectation is that data is returned, and EOF is
signalled afterwards.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>