modules: make module support standards-compliant
C++ doesn't allow #ifdef tricks with modules. Specifically, the module
keywords can't be guarded with #ifdef.

Here, we change the approach towards module support. It is heavily
inspired by libstdc++ [1]. In the new approach, instead of two libraries,
seastar and seastar-module, one without and one with module support, there
is only one library. Module support comes in the form of a .cppm file
that declares the module and exports the symbols.

Changes in the patch:

1. remove SEASTAR_MODULE_EXPORT* defines; the export declarations are
   moved to seastar.cppm (causing some duplication)
2. remove the SEASTAR_MODULE define and unifdef the code it guards. There
   is a single complication pass rather than two.
3. The utils/modules.hh header is removed, since it no longer defines
   anything.
4. seastar.cc is renamed to seastar.cppm so the compiler understands it's
   a module definition. It now only includes headers that contain public
   symbols. The symbols to be exported are listed explicitly.
5. The cmake seastar library now conditionally includes seastar.cppm.
6. Minor cmake changes for hello-cxx-module to make it build.

[1] https://github.com/gcc-mirror/gcc/blob/master/libstdc%2B%2B-v3/src/c%2B%2B23/std.cc.in

Tested with clang 21.

Fixes https://github.com/scylladb/seastar/issues/3091

Closes scylladb/seastar#3101